Clean Up
Basic Cleanup of service by splitting Job into Job & JobER
