+++ src/app/shared/services/job.service.ts
@@ -5,7 +5,6 @@ import 'firebase/firestore';
 import { Observable } from 'rxjs';
 import { Time } from '@angular/common';
 import { ErrandRunner } from '../models/ErrandRunner';
-import { isNullOrUndefined } from 'util';
 
 @Injectable({
   providedIn: 'root'
@@ -243,56 +242,6 @@ export class JobService {
 
     })
   }
-  getAllErrands(): Observable<any> {
-    return new Observable(observer => {
-      // Read collection '/JobsAvailable'
-      firebase.firestore().collection('JobsAvailable').onSnapshot(collection => {
-        let array = [];
-        collection.forEach(doc => {
-
-          // Add jobs into array if there's no error
-          try {
-            let loan = new Job(doc.data().errandname, doc.data().category, doc.data().status, doc.data().client, doc.data().date.toDate(), doc.data().description, doc.data().time, doc.id);
-            array.push(loan);
-
-          } catch (error) { }
-
-        });
-        observer.next(array);
-      });
-    });
-  }
-
-  getAllErrandsExcept(id: string): Observable<any> {
-    return new Observable(observer => {
-      // Read collection '/JobsAvailable'
-      const ref = firebase.firestore().collection('JobsAvailable')
-      ref.where('client', '!=', id).onSnapshot(collection => {
-        let array = [];
-        collection.forEach(doc => {
-
-          // Add jobs into array if there's no error
-          try {
-            const docRef = ref.doc(doc.id)
-            docRef.collection('Applicants').get().then(sdoc => {
-              let applied = null
-              docRef.collection('Applicants').doc(id).get().then((docSnapshot)=>{
-                if(docSnapshot.exists){
-                  return;
-                }
-                else{
-                  let loan = new Job(doc.data().errandname, doc.data().category, doc.data().status, doc.data().client, doc.data().date.toDate(), doc.data().description, doc.data().time, doc.id);
-                  array.push(loan);
-                }
-              })
-            })
-          } catch (error) { }
-
-        });
-        observer.next(array);
-      });
-    });
-  }
 
   applyjobs(id: string, email: string, dateapplied: Date) {
     return firebase.firestore().collection('JobsAvailable').doc(id).collection('Applicants').get().then(collection => {
@@ -374,75 +323,6 @@ export class JobService {
         jobref.delete();
       }
     })
-  }
-
-  getAllErrandsApplied(id: string): Observable<any> {
-    return new Observable(observer => {
-      // Read collection '/JobsAvailable'
-      const ref = firebase.firestore().collection('JobsAvailable')
-      ref.onSnapshot(collection => {
-        let array = [];
-        
-        collection.forEach(doc => {
-          // Add jobs into array if there's no error
-          try {
-            const docRef = ref.doc(doc.id)
-            docRef.collection('Applicants').get().then(sdoc => {
-              let applied = null
-              sdoc.forEach(ssdoc => {
-                if (ssdoc.id === id) {
-                  applied = true
-                }
-                else{
-                  applied = false
-                }
-
-                if (applied === true) {
-                  let loan = new Job(doc.data().errandname, doc.data().category, doc.data().status, doc.data().client, doc.data().date.toDate(), doc.data().description, doc.data().time, doc.id);
-                  array.push(loan);
-                }
-              })
-            })
-          } catch (error) { }
-
-        });
-        observer.next(array);
-      });
-    });
-  }
-
-  getAllErrandsAccepted(id: string): Observable<any> {
-    return new Observable(observer => {
-      // Read collection '/JobsAccepted'
-      const ref = firebase.firestore().collection('JobsAccepted')
-      ref.onSnapshot(collection => {
-        let array = [];
-        let applied = false
-        collection.forEach(doc => {
-          // Add jobs into array if there's no error
-          try {
-            const docRef = ref.doc(doc.id)
-            docRef.collection('Applicant').get().then(sdoc => {
-              sdoc.forEach(ssdoc => {
-                if (ssdoc.id === id) {
-                  applied = true
-                }
-                else {
-                  applied = false
-                }
-
-                if (applied === true) {
-                  let loan = new Job(doc.data().errandname, doc.data().category, doc.data().status, doc.data().client, doc.data().date.toDate(), doc.data().description, doc.data().time, doc.id);
-                  array.push(loan);
-                }
-              })
-            })
-          } catch (error) { }
-
-        });
-        observer.next(array);
-      });
-    });
   }
 
   getConfirmedJobsByClient(client: string): Observable<any> {
